+++ DemoFramework/FslDataBinding/Base/include/FslDataBinding/Base/DataSourceFlags.hpp
@@ -33,11 +33,9 @@
 
 #include <FslBase/BasicTypes.hpp>
 
-#include <cstdint>
-
 namespace Fsl::DataBinding
 {
-  enum class DataSourceFlags : uint8_t
+  enum class DataSourceFlags : uint32_t
   {
     NoFlags = 0,
     Observable = 0x01,
