+++ DemoFramework/FslDemoApp/Base/include/FslDemoApp/Base/CustomDemoAppConfig.hpp
@@ -33,14 +33,12 @@
 
 #include <FslBase/ITag.hpp>
 #include <FslDemoService/Graphics/ColorSpaceType.hpp>
-#include <cstdint>
 #include <memory>
 #include <utility>
 
-
 namespace Fsl
 {
-  enum class CustomDemoAppConfigRestartFlags : uint8_t
+  enum class CustomDemoAppConfigRestartFlags : uint32_t
   {
     //! No flags set, so no restart
     Never = 0x00,
