+++ DemoFramework/FslGraphics/include/FslGraphics/TextureFlags.hpp
@@ -33,11 +33,9 @@
 
 #include <FslBase/BasicTypes.hpp>
 
-#include <cstdint>
-
 namespace Fsl
 {
-  enum class TextureFlags : uint8_t
+  enum class TextureFlags : uint32_t
   {
     // A empty flag
     NotDefined = 0x00,
